requirements.txt
argparse
straight.plugin==1.4.0-post-1
cssselect
pyquery==1.2.9
numpy
biopython>=1.65
helperlibs
pysvg
pyExcelerator
backports.lzma
networkx==1.11
python-louvain==0.5
bcbio-gff>=0.6.2
